requirements.txt
lxml==3.7.2
xmltodict==0.10.2
Flask
flask-mongoengine
